03.ClientPlatform/04.Animation/src/main/webapp/app/entities/you-tube-video/route/you-tube-video-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IYouTubeVideo } from '../you-tube-video.model';
import { YouTubeVideoService } from '../service/you-tube-video.service';

import youTubeVideoResolve from './you-tube-video-routing-resolve.service';

describe('YouTubeVideo routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let service: YouTubeVideoService;
  let resultYouTubeVideo: IYouTubeVideo | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    service = TestBed.inject(YouTubeVideoService);
    resultYouTubeVideo = undefined;
  });

  describe('resolve', () => {
    it('should return IYouTubeVideo returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      TestBed.runInInjectionContext(() => {
        youTubeVideoResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultYouTubeVideo = result;
          },
        });
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultYouTubeVideo).toEqual({ id: 123 });
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      TestBed.runInInjectionContext(() => {
        youTubeVideoResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultYouTubeVideo = result;
          },
        });
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultYouTubeVideo).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<IYouTubeVideo>({ body: null })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      TestBed.runInInjectionContext(() => {
        youTubeVideoResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultYouTubeVideo = result;
          },
        });
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultYouTubeVideo).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
